Trade Menu Refactoring
Changed the way the trade menu works by adding new functions which control the input of the user. By doing this it is possible to change both the players and stores inventory in the same function.

Changed the gamestate enum to reflect the changes as two seperate states aren't needed to change the inventories.

TODO: Fix the exit condition when selecting an item